model-zoo: fix all 18 smoke tests on Casper V100
Key fixes across models and infrastructure:

Models:
- aurora: handle 5D (B,C,T,H,W) input, pad H=180→181 after Aurora crop,
  bypass double-normalisation by setting identity surf_stats
- aurora/normalisation: fall back to identity (loc=0, scale=1) for
  pressure levels not in Aurora's built-in table
- pangu: handle 5D input, add unsqueeze(2) to return 5D output
- aifs: absorb parser-injected `levels` kwarg via signature filter
- climax: pad H/W to patch_size multiple, crop output to original size
- corrdiff: pad H/W to 2^n_downs boundary, crop output
- fourcastnet3: pad H/W to 2^n_stages boundary, crop output; fix skip
  connection crop to min(encoder, decoder) size; add frames param
- graphcast: fix scatter dtype mismatch (use edge_feat.dtype for agg)
- swinrnn: align padding to patch_size * window_size * 2^(n_stages-1)
- wxformer_next: fix decoder skip connections with F.interpolate;
  fix residual channel mismatch with F.pad

Configs:
- Add levels: 18 to data+model sections for arches, corrdiff,
  fuxi_ens, itransformer, mambavision
- Add levels: 18 to data section for aifs, aurora, pangu
- sfno: reduce n_modes 90->46, add use_harmonics: False
- fourcastnet3: disable amp (half-precision FFT fails on non-power-of-2)
- nextgen_wxformer: fix local/global window sizes to per-stage lists
  that divide token dims at every stage

Infrastructure:
- parser: make model-level levels check optional (.get with None)
- weighted_loss + metrics (3 sites): fall back through
  model.levels -> model.frames -> data.levels for variable list

+++ credit/models/graphcast/graphcast.py
@@ -139,7 +139,7 @@ class MessagePassingLayer(nn.Module):
         edge_feat = self.edge_mlp(src_feat, dst_feat, edge_feat)
 
         # aggregate (sum) edges to destination nodes
-        agg = torch.zeros(B, N, edge_feat.shape[-1], device=node_feat.device, dtype=node_feat.dtype)
+        agg = torch.zeros(B, N, edge_feat.shape[-1], device=node_feat.device, dtype=edge_feat.dtype)
         agg.scatter_add_(1, dst[None, :, None].expand(B, -1, edge_feat.shape[-1]), edge_feat)
 
         # node update
